add lock file with first dependency
We have built this in root with `yarn`.
Thanks to workspaces yarn understands that pacakges have deps
that need to be added to the root worspace file.
Same when in a packages folder, E.G. in types/ yarn understands that's a
single package with its own "context" so to say.
In fact, executing `yarn build` it recognizes the script in
its `.package.json` and executes it accordingly.